requirements.txt
vllm
sentence-transformers
faiss-cpu  # or faiss-gpu if you're using GPU
numpy
torch 